ESC-970 During the analysis of ESC-970 it was found that Agent is allocating a lot of memory. From the charts shared by support, this correlated to the increasing number of containers on the host. Zipper's memdump analysis showed memory allocations happening in load_k8s_audit_policy_v2(). On further analysis, it was found this function is called for every new container, if there is one or more k8s audit policy pused from the backend. In case there is a lot of container churn, this function will be called everytime a new container comes up and allocate extraneous memory. This memory is freed on receipt of policies_v2 message (default frequency every 10 mins). But during these 10 mins a huge amount of memory can be allocated corresponding to the number of containers coming up.
This function load_k8s_audit_policy_v2() is not dealing with the new
container id so it does not need to be called for every new container.
It will be called every time policies_v2 is received every 10 mins
anyways. This change shows a good amount of memory saving.
A container churn test was done as follows:
start 45 containers,, let them run for 2 mins, remove all containers and
then wait for 4 mins. repeat this cycle for 5-6 times and watch the
memory.
Without the fix, dragent memory grew from ~170MB to > 500MB
With the fix, memory grew from ~140MB to ~178MB only