[SMAGENT-2878] Part 2: Remove capture job stuff from sinsp worker (#2750)
the ultimate goal of this task (2878) is to pare apart sinsp_worker to get independence of secure and the analyzer (and compliance manager and capture job handler) to allow the infras to be completely independent, allowing them to run without the others trivially or in parallel.

In order to kill sinsp_worker, though, we have to start moving much of the things that have sort of been orphaned there over the years to where they belong. This change accomplishes much of that with respect to the capture_job_handler, namely, the infra around dealing with capture job requests from the backend, which for no reason, made a loop through the sinsp_worker (you'll note the capture job handler ALREADY IS THREADSAFE afaict).

To give some background, there are 3 ways that dumps are triggered:

on reception of a signal. this signal handling is moved from the sinsp_worker loop to the main dragent loop, where signal handling almost surely should be anyway, and is then directly handed to the new api on the capture job infra.
on reception of a dump start request from the backend. this is now directly handled by the capture_job_handler (well, a semi-extraneous friend), instead of being bufferred by sinsp_worker for no reason.
when a security policy needs it. this also goes directly to capture job.
The most difficult part of this change had to do with how error codes are dealt with. I don't believe it was clean before, and I also don't believe it is now. There are some cases where the infra returned status, expecting the caller to do things, some places where it automatically sent status to the backend, and some cases where these are dropped on the floor. I tried my best to make as minimal change in this area as possible in order to maintain bug-for-bug compatibility, but may not have succeeded. I believe, though. One odd case, for example, is how the errors are dealt with in the call from security manager. It seems we only propagate the failure to security manager when a capture fails to start, not if it fails later on. Given it's alreayd broken, seemingly, I didn't make much effort to make it less broken.