[SMAGENT-2878] Part 3: Remove policies_v2 handler from sinsp_worker #2752
the ultimate goal of this task (2878) is to pare apart sinsp_worker to get independence of secure and the analyzer (and compliance manager and capture job handler) to allow the infras to be completely independent, allowing them to run without the others trivially or in parallel.

In order to kill sinsp_worker, though, we have to start moving much of the things that have sort of been orphaned there over the years to where they belong. This change accomplishes much of that with respect to the policies_v2 message handler.

This change is pretty straightforward, and the agentino already goes straight to the security manager instead of depending on sinsp_worker, so this should "just work" tm.