[SMAGENT-3244] ut: handle ERR_DUPLICATE_AGENT
ERR_DUPLICATE_AGENT is an error by which the backend indicates
its conviction about an agent with the same machine id being
already connected. The nature of this error is likely transient
and might be attributable to Sysdig.
So when received during handshake, we should disconnect
and slighly back off (or just keep the existing backoff time)
to keep downtime to a minimum.

ERR_INVALID_CUSTOMER_KEY and ERR_CONN_LIMIT on the other hand
likely indicate a problem on the customer's licensing side
which won't fix itself.
So in that case we want to back off to avoid pounding the backend.

This test exercises the following scenario during handshake
1. Send ERR_INVALID_CUSTOMER_KEY -> disconnect and set backoff to 1s, for starters
2. Send ERR_DUPLICATE_AGENT -> disconnect and keep backoff to 1s
3. Send ERR_CONN_LIMIT -> disconnect and double backoff time to 2s
4. Send ERR_DUPLICATE_AGENT -> disconnect and keep backoff to 2s