[SMAGENT-3259] Set available to true if we successfully load JMX and receive a valid address via loadManagementAgent() (#2826)
This change is simple, but it makes a significant difference. Previously, when loadManagementAgent() successfully loaded the JMX agent into the monitored process, it wouldn't take effect in sdjagent. Instead, we'd effectively ignore it and then try to load the JMX info from the command-line arguments of the process. Basically, loadManagementAgent() was a no-op in terms of sdjagent.

This change will make it so that if loadManagementAgent() is successful, we'll use the JMX info we get from it. From looking at the git blame for this file, it appears this behavior has always been in sdjagent, going back 5 years to that function's first existence. I don't know if there's a reason for this or if this is just a bug that has escaped for a really long time.

I tested this on the Websphere Liberty Docker image which was now working and also tested that it was working with a couple other applications in different JVMs: Cassandra and Tomcat.